new hook & some qof updates
