+++ debugNote.md
@@ -1,0 +1,9 @@
+# 17/11/2024 bugs fixed:
+1. исправлена ошибка с неправильным указанием параметров запроса:
+    вместо параметра dt для history запроса был указан параметр days необходимый для forecast запроса.
+2. добавлена проверка типа запросе перед форматированием history запроса,
+    который вызывал ошибку из-зи отсутствия необходимых полей ответа. 
+
+# 20/11/2024 updates:
+1. улучшена оптимизация приложения через React.memo
+2. добавлен роутинг
