+++ src/pages/Main.jsx
@@ -1,0 +1,9 @@
+import { memo } from 'react'
+
+const Main = memo(() => {
+	console.log('Main')
+
+	return <div>вшм</div>
+})
+
+export default Main
